fix: fixed unit tests & removed integration tests
