+++ .github/workflows/checks.yml
@@ -35,25 +35,6 @@ jobs:
 
       - name: Npm run test
         run: npm run test
-  integration-test:
-    runs-on: ubuntu-latest
-    continue-on-error: true
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-
-      - name: Setup Node.js
-        uses: actions/setup-node@v3
-        with:
-          node-version: 18
-
-      - name: Install dependencies
-        run: npm ci
-
-      - name: Npm run test:integration
-        run: npm run test:integration
-        env:
-          TEST_SECRETS: ${{secrets.TEST_ENV_FILE}}
   call-security-scanners-workflow:
     uses: affinidi/pipeline-security/.github/workflows/security-scanners.yml@feat/check-inherit
     with:
