+++ package.json
@@ -20,9 +20,7 @@
     "build": "shx rm -rf dist && tsc -b",
     "lint": "eslint . --ext .ts",
     "lint:fix": "eslint . --ext .ts --fix",
-    "test": "mocha \"test/unit/**/*.test.ts\"",
-    "test:integration": "mocha --config test/integration/.mocharc.json \"test/integration/**/*.test.ts\"",
-    "checkLicenses": "license-checker --production --failOn 'AGPL 3.0;LGPL 2.1'",
+    "test": "mocha \"test/**/*.test.ts\"",
     "generate:readme": "npm run build && oclif readme",
     "generate:api": "node scripts/generate-api.cjs"
   },
@@ -58,15 +56,12 @@
     "axios": "^1.4.0",
     "chalk": "^4.1.2",
     "conf": "^10.2.0",
-    "cookie-parser": "^1.4.6",
     "degit": "^2.8.4",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "http-status-codes": "^2.2.0",
     "jose": "^4.14.6",
-    "jsonwebtoken": "^9.0.2",
     "open": "^8.4.1",
-    "simple-oauth2": "^5.0.0",
     "tcp-port-used": "^1.0.2",
     "zod": "^3.22.2",
     "zod-validation-error": "^1.5.0"
@@ -76,14 +71,10 @@
     "@semantic-release/git": "^10.0.1",
     "@semantic-release/npm": "^11.0.0",
     "@types/chai": "4.3.5",
-    "@types/cookie-parser": "^1.4.5",
     "@types/degit": "^2.8.4",
     "@types/express": "^4.17.17",
-    "@types/fs-extra": "^11.0.2",
-    "@types/jsonwebtoken": "^9.0.3",
     "@types/mocha": "10.0.1",
     "@types/node": "^18.11.19",
-    "@types/simple-oauth2": "^5.0.4",
     "@types/tcp-port-used": "^1.0.1",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^5.59.9",
@@ -95,12 +86,10 @@
     "eslint-plugin-import": "^2.28.1",
     "eslint-plugin-mocha": "^10.1.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "license-checker": "25.0.1",
     "mocha": "^10.2.0",
     "nyc": "15.1.0",
     "oclif": "^3.9.0",
     "prettier": "^2.7.1",
-    "qs": "^6.11.2",
     "semantic-release": "^21.1.2",
     "shx": "^0.3.4",
     "swagger-typescript-api": "^12.0.4",
@@ -140,9 +129,6 @@
     },
     "warn-if-update-available": {
       "timeoutInDays": 7
-    },
-    "hooks": {
-      "check": "./dist/hooks/check/check-version"
     }
   },
   "engines": {
