+++ src/commands/iam/update-policies.ts
@@ -5,12 +5,12 @@ import { CLIError } from '@oclif/core/lib/errors'
 import chalk from 'chalk'
 import { z } from 'zod'
 import { BaseCommand, PrincipalTypes } from '../../common'
-import { promptRequiredParameters } from '../../common/prompts'
 import { giveFlagInputErrorMessage } from '../../common/error-messages'
+import { promptRequiredParameters } from '../../common/prompts'
 import { INPUT_LIMIT, validateInputLength } from '../../common/validators'
+import { bffService } from '../../services/affinidi/bff-service'
 import { iamService } from '../../services/affinidi/iam'
 import { PolicyDto } from '../../services/affinidi/iam/iam.api'
-import { bffService } from '../../services/affinidi/bff-service'
 
 export class UpdatePolicies extends BaseCommand<typeof UpdatePolicies> {
   static summary = 'Updates the policies of a principal (user or token) in the active project'
