+++ src/services/affinidi/vp-adapter/index.ts
@@ -10,9 +10,9 @@ import {
   GroupsList,
   GroupUserMappingsList,
 } from './vp-adapter.api'
-import { handleServiceError } from '../errors'
 import { config } from '../../env-config'
 import { getBFFHeaders } from '../bff-service'
+import { handleServiceError } from '../errors'
 
 export const VPA_SERVICE = 'vp-adapter'
