+++ test/commands/iam/principals.test.ts
@@ -1,11 +1,9 @@
 import { expect, test } from '@oclif/test'
-import { configService } from '../../../../src/services'
-import { clientSDK } from '../../../../src/services/affinidi'
-import { IAM_URL } from '../../../../src/services/urls'
+import { config } from '../../../src/services/env-config'
 
 // TODO: Extract the mocked data in a shared folder, remove mock duplicates if any.
 
-const testUserId = '38efcc70-bbe1-457a-a6c7-b29ad9913648'
+const IAM_URL = `${config.bffHost}/iam`
 const principalId = 'b2ce7675-5418-4058-b973-d254270de2d4'
 const principalType = 'machine_user'
 const listPrincipalsApiResponse = {
@@ -31,13 +29,6 @@ const listPrincipalsApiResponse = {
 
 describe('iam: commands', function () {
   describe('add-principal', function () {
-    beforeEach(function () {
-      configService.create(testUserId)
-    })
-    afterEach(function () {
-      configService.clear()
-      clientSDK.config.clear()
-    })
     test
       .nock(IAM_URL, (api) => api.post('/v1/projects/principals').reply(200))
       .stdout()
@@ -46,20 +37,12 @@ describe('iam: commands', function () {
   })
 
   describe('list-principals', function () {
-    beforeEach(function () {
-      configService.create(testUserId)
-    })
-    afterEach(function () {
-      configService.clear()
-      clientSDK.config.clear()
-    })
-
     test
       .nock(IAM_URL, (api) => api.get('/v1/projects/principals').reply(200, listPrincipalsApiResponse))
       .stdout()
       .command(['iam list-principals'])
       .it('Should return list of projects', (ctx) => {
-        const response: any = JSON.parse(ctx.stdout)
+        const response = JSON.parse(ctx.stdout)
         expect(response).to.have.a.property('records')
         expect(response?.records).to.be.instanceOf(Array)
         expect(response?.records).to.have.length.greaterThanOrEqual(1)
@@ -75,14 +58,6 @@ describe('iam: commands', function () {
   })
 
   describe('remove-principal', function () {
-    beforeEach(function () {
-      configService.create(testUserId)
-    })
-    afterEach(function () {
-      configService.clear()
-      clientSDK.config.clear()
-    })
-
     test
       .nock(IAM_URL, (api) =>
         api.delete(`/v1/projects/principals/${principalId}?principalType=${principalType}`).reply(200),
