+++ test/helpers/init.js
@@ -3,7 +3,7 @@ const path = require('path')
 process.env.TS_NODE_PROJECT = path.resolve('tsconfig-test.json')
 process.env.NODE_ENV = 'test'
 // TODO remove or use carefully
-process.env.AFFINIDI_CLI_ENVIRONMENT = 'test'
+process.env.AFFINIDI_CLI_ENVIRONMENT = 'local'
 // END OF TODO remove or use carefully
 
 global.oclif = global.oclif || {}
