--- test/integration/commands/iam/data.ts
@@ -1,4 +1,0 @@
-export const data = {
-  principalId: 'b2ce7675-5418-4058-b973-d254270de2d0',
-  principalType: 'machine_user',
-}
