--- test/integration/commands/login/data.ts
@@ -1,11 +1,0 @@
-export const data = {
-  existingGroupName: 'existing_group_name',
-  newGroupName: 'new_group_name',
-  newUserSub: 'did:elem:12345',
-  invalidGroupName: 'group name',
-  nonExistentGroupName: '12345',
-  userGroupMappingId: 'd07317870727e7ab86120ddfaee77a2ba371720f14975df3d8d52c818d6481ea',
-  configId: 'd1b0943a965579d39980f084840a6882',
-  deleteConfigId: 'abc',
-  configName: 'test-config',
-}
