--- test/unit/services/constants.ts
@@ -1,28 +1,0 @@
-export const PROJECTS = [
-  {
-    id: '1',
-    name: 'Project1',
-  },
-  {
-    id: '2',
-    name: 'Project2',
-  },
-  {
-    id: '3',
-    name: 'Project3',
-  },
-  {
-    id: '4',
-    name: 'Project4',
-  },
-]
-
-export const PROJECT_1 = {
-  id: '1',
-  name: 'Project1',
-}
-
-export const PROJECT_2 = {
-  id: '2',
-  name: 'Project2',
-}
